Extract ions.py from _compute/runtime.py
Replace the re-export shim in braincell/_compute/ions.py with the real
bodies of the 14 runtime-ion symbols, moved verbatim out of runtime.py:
ion instance collection and instantiation, alias/species/family naming,
runtime param introspection and normalization, the broadcast/scatter
buffer algebra, and per-layout ion synchronization.

Pure relocation; all 14 bodies are byte-identical to their previous text.
ions.py imports nothing from runtime.py: it depends downward on layouts
for MechanismLayout and _constant_quantity_value, outward on the public
braincell.ion and braincell.mech packages, and refers to CellRuntimeState
only through a TYPE_CHECKING import of .state.

runtime.py drops the now-orphaned inspect import and the unused
DynamicNernstIon/FixedIon/InitNernstIon names, and re-exports all 14
moved symbols so its attribute surface is unchanged.
